segmentaion/requirements.txt
opencv-python
pyrebase4
sseclient
python_jwt
gcloud
pycryptodome
requests-toolbelt
pillow